# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "瞬捷云 captcha-api 验证码服务api",
        "title": "Swagger Example API",
        "termsOfService": "https://www.shunjiecloud.com",
        "contact": {
            "name": "zhoushengjie",
            "url": "https://www.shunjiecloud.com",
            "email": "zhou_shengjie@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "api.shunjiecloud.com",
    "basePath": "/",
    "paths": {
        "/captcha/v1/captcha": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new pet to the store",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/captcha/v1/captcha/{filename}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/png"
                ],
                "summary": "Add a new pet to the store",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证码id",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证码图片"
                    },
                    "400": {
                        "description": "404 page not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}